Conditional service startup for contrail components
To support upgrades, we need a way to control of services
start-up after components deployment. If a flag variable is found
in Hiera, manifests do not ensure services to be started.
Also, ubuntu-specific tweaks were added to manifests to prevent
automation service starts and restarts after package installation
or upgrades.

Change-Id: I13c607078fc2fd4f26a90665a217b1c40288ed04